moving files to correct locations and deleting old/out of date files
